move mockData inside backend so that
I don't need to change de buildBackend script
